% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\title{dotplot}
\usage{
dotplot(lona, lata, map = "ALL", eez = TRUE, grid = FALSE)
}
\arguments{
\item{lona}{Input the longitude.}

\item{lata}{Input the latitude.}

\item{map}{default is "ALL", Other possible options is "PAC", "IND" and "ATL".}

\item{eez}{default is TRUE, when TRUE show the EEZ map.}

\item{grid}{default is FALSE, when TRUE show the 5 degree grid.}
}
\value{
the plot of \code{lona} and \code{lata}.
}
\description{
This function allows you to draw data distribution geographically with EEZ map Function from a numeric vector.
}
\examples{
dotplot(141,23)
}
