% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eez_rg}
\alias{eez_rg}
\title{Eez Coefficients}
\format{
A data.frame with 2 variables: \code{geneName}, \code{coef}
}
\usage{
eez_rg
}
\description{
Predictor feature coefficients as published in paper.
}
\keyword{datasets}
