% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idfcode.R
\name{idfcode}
\alias{idfcode}
\title{idfcode}
\usage{
idfcode(lon, lat)
}
\arguments{
\item{lon}{Input the longitude.}

\item{lat}{Input the latitude.}
}
\description{
This function allows you to convert the location to 4 digital code
}
\examples{
idfcode(22,-5)
}
