% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idfeez.R
\name{idfeez}
\alias{idfeez}
\title{idfeez}
\usage{
idfeez(lon, lat, ac = TRUE)
}
\arguments{
\item{lon}{Input the longitude.}

\item{lat}{Input the latitude.}

\item{ac}{logical. If \code{TRUE} will return full name of EEZ.}
}
\description{
This function allows you to identify location in which EEZ from a numeric vector.
}
\examples{
idfeez(141,23)
}
