% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idfland.R
\name{idfland}
\alias{idfland}
\title{idfland}
\usage{
idfland(lon, lat)
}
\arguments{
\item{lon}{Input the longitude.}

\item{lat}{Input the latitude.}
}
\description{
This function allows you to identify location in which land or ocean.
}
\examples{
idfland(22,-5)
}
