% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idfport.R
\name{idfport}
\alias{idfport}
\title{idfport}
\usage{
idfport(lon, lat)
}
\arguments{
\item{lon}{Input the longitude.}

\item{lat}{Input the latitude.}
}
\description{
This function allows you to identify port name from a numeric vector.
}
\examples{
idfport(121.8006,25.14065)
}
