% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanic.R
\name{oceanic}
\alias{oceanic}
\title{oceanic}
\usage{
oceanic(lon, lat)
}
\arguments{
\item{lon}{Input the longitude.}

\item{lat}{Input the latitude.}
}
\value{
the ocean of \code{lon} and \code{lat}.
}
\description{
Return The Pacific Ocean(PAC), Indian Ocean(IND) or Atlantic Ocean(ATL) of your coordinate.
}
\examples{
oceanic(125,20)
}
