% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixtytoten.R
\name{sixtytoten}
\alias{sixtytoten}
\title{sixtytoten}
\usage{
sixtytoten(num)
}
\arguments{
\item{num}{Input a value of longitude or latitude.}
}
\description{
This function allows you to transfer the coordinate system from sexagesimal to decimal
}
\examples{
sixtytoten(121.49)
}
